Fix options page in VS2017 15.9.12
While the RustLanguageExtension.dll assembly is loaded fine for
intellisense, syntax highlighting, etc. - the Tools > Options > Rust
page fails to load the appropriate assembly.  This might be related to
VS2017 breaking changes[1], but I'm not 100% sure what's going on.

I'm not even sure this isn't a bug in VS, given that only the options
page is unable to load the assembly, but at least we have a fix [2].

[1] https://docs.microsoft.com/en-us/visualstudio/extensibility/breaking-changes-2017?view=vs-2019#change-global-assembly-cache-gac
[2] https://social.msdn.microsoft.com/Forums/en-US/39ca5319-96af-4de7-b2d1-4d28154321b4/vsix-options-dialog-does-not-always-load-in-vsix-extension?forum=vsx#b209ab3b-6f0d-493f-a322-c25e4ed6b82e